requirements.txt
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.11.12
click==8.3.0
ConfigArgParse==1.7.1
contourpy==1.3.3
cycler==0.12.1
dash==3.2.0
databricks==0.2
dotenv==0.9.9
easydict==1.13
fastapi==0.121.1
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
h11==0.16.0
h5py==3.15.1
hydra-core==1.3.2
idna==3.11
importlib_metadata==8.7.0
ipywidgets==8.1.7
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lightning==2.5.6
lightning-utilities==0.15.2
MarkupSafe==2.1.5
matplotlib==3.10.7
mlflow==3.6.0
mpmath==1.3.0
multidict==6.7.0
narwhals==2.8.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
numpy==2.2.0
omegaconf==2.3.0
open3d==0.19.0
opencv-python==4.12.0.88
opentelemetry-api==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
plotly==6.3.1
propcache==0.4.1
protobuf==6.33.0
pydantic==2.12.4
pydantic_core==2.41.5
pyparsing==3.2.5
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytorch-lightning==2.5.6
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
retrying==1.4.2
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.49.3
sympy==1.14.0
tensorboardX==2.6.4
termcolor==3.1.0
threadpoolctl==3.6.0
timm==1.0.20
torchmetrics==1.8.2
tqdm==4.67.1
transforms3d==0.4.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
Werkzeug==3.1.3
yarl==1.22.0
zipp==3.23.0
